+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Version 3.2.1 (to be released)
 
 * Fixed JToggleButtonSelectedProperty value
+* Fixed initial value for ButtonPressedProperty and JTextComponentTestProperty
 
 # Version 3.2.0